requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic
passlib[bcrypt]
python-jose
python-dotenv
pydantic[email]
pandas
email-validator
python-multipart
faiss-cpu
openai>=1.3.0
passlib>=1.7.4
bcrypt>=4.0.0